create Abins2 algorithm, not subscribed by default
